ml/requirements-ml.txt
# Machine Learning Dependencies for Electricity Price Forecasting

# Deep Learning
tensorflow>=2.15.0,<3.0
keras>=2.15.0
tensorboard>=2.15.0

# ML Frameworks
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.3.0

# Hyperparameter Optimization
optuna>=3.5.0

# Optimization
pulp>=2.7.0
scipy>=1.12.0

# Data Processing
numpy>=1.26.0,<2.0
pandas>=2.1.0
holidays>=0.40

# Time Series
statsmodels>=0.14.0
prophet>=1.1.5

# Feature Engineering
category-encoders>=2.6.0

# Model Serialization
joblib>=1.3.0
pickle5>=0.0.12

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Monitoring
mlflow>=2.9.0
wandb>=0.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
